requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.3
python-multipart>=0.0.9

# Database
supabase>=2.4.0

# Task queue
celery>=5.3.6
redis>=5.0.3

# LLM parsing
anthropic>=0.25.0

# SMS
twilio>=9.0.0

# Google APIs (email + sheets + drive)
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.126.0

# HTTP / scraping
httpx>=0.27.0
beautifulsoup4>=4.12.3
playwright>=1.43.0

# PDF / OCR
pdfplumber>=0.11.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.3.0

# Config / utilities
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.1

# Testing
pytest>=8.1.0
pytest-asyncio>=0.23.6
httpx  # also used for test client
